# Update Kubernetes from 1.27.3 to 1.30.4