--- a/examples/go.mod
+++ b/examples/go.mod
@@ -1,6 +1,6 @@
 module sigs.k8s.io/kube-scheduler-wasm-extension/examples
 
-go 1.20
+go 1.22.0
 
 require (
 	sigs.k8s.io/kube-scheduler-wasm-extension/guest v0.0.0-00010101000000-000000000000