--- a/examples/imagelocality/go.mod
+++ b/examples/imagelocality/go.mod
@@ -1,14 +1,10 @@
 module sigs.k8s.io/kube-scheduler-wasm-extension/examples/imagelocality
 
-go 1.20
+go 1.22.0
+
+require sigs.k8s.io/kube-scheduler-wasm-extension/guest v0.0.0-00010101000000-000000000000
 
 require (
-	github.com/wasilibs/nottinygc v0.7.1
-	sigs.k8s.io/kube-scheduler-wasm-extension/guest v0.0.0-00010101000000-000000000000
-)
-
-require (
-	github.com/magefile/mage v1.14.0 // indirect
 	google.golang.org/protobuf v1.30.0 // indirect
 	sigs.k8s.io/kube-scheduler-wasm-extension/kubernetes/proto v0.0.0-00010101000000-000000000000 // indirect
 )