--- a/guest/testdata/go.mod
+++ b/guest/testdata/go.mod
@@ -1,6 +1,6 @@
 module sigs.k8s.io/kube-scheduler-wasm-extension/guest/testdata
 
-go 1.20
+go 1.22.0
 
 require (
 	sigs.k8s.io/kube-scheduler-wasm-extension/guest v0.0.0-00010101000000-000000000000