--- a/internal/e2e/scheduler_perf/nodenumber/plugin.go
+++ b/internal/e2e/scheduler_perf/nodenumber/plugin.go
@@ -55,7 +55,7 @@ func (s *preScoreState) Clone() framework.StateData {
 	return s
 }
 
-func (pl *NodeNumber) PreScore(ctx context.Context, state *framework.CycleState, pod *v1.Pod, nodes []*v1.Node) *framework.Status {
+func (pl *NodeNumber) PreScore(ctx context.Context, state *framework.CycleState, pod *v1.Pod, nodes []*framework.NodeInfo) *framework.Status {
 	podNameLastChar := pod.Name[len(pod.Name)-1:]
 	podnum, err := strconv.Atoi(podNameLastChar)
 	if err != nil {
@@ -120,7 +120,7 @@ func (pl *NodeNumber) ScoreExtensions() framework.ScoreExtensions {
 }
 
 // New initializes a new plugin and returns it.
-func New(arg runtime.Object, h framework.Handle) (framework.Plugin, error) {
+func New(ctx context.Context, arg runtime.Object, h framework.Handle) (framework.Plugin, error) {
 	typedArg := NodeNumberArgs{Reverse: false}
 	if arg != nil {
 		err := frameworkruntime.DecodeInto(arg, &typedArg)