--- a/kubernetes/proto/go.mod
+++ b/kubernetes/proto/go.mod
@@ -1,5 +1,5 @@
 module sigs.k8s.io/kube-scheduler-wasm-extension/kubernetes/proto
 
-go 1.20
+go 1.22.0
 
 require google.golang.org/protobuf v1.30.0