… examples/advanced/go.sum
… examples/go.sum
… examples/imagelocality/go.sum
… guest/testdata/go.sum
… internal/e2e/go.sum
… internal/e2e/guest/go.sum
… kubernetes/proto/api/generated.pb.go
… kubernetes/proto/api/generated_vtproto.pb.go
… kubernetes/proto/instr/generated.pb.go
… kubernetes/proto/instr/generated_vtproto.pb.go
… kubernetes/proto/meta/generated.pb.go
… kubernetes/proto/meta/generated_vtproto.pb.go
… kubernetes/proto/resource/generated.pb.go
… kubernetes/proto/resource/generated_vtproto.pb.go
… kubernetes/proto/runtime/generated.pb.go
… kubernetes/proto/runtime/generated_vtproto.pb.go
… kubernetes/proto/schema/generated.pb.go
… kubernetes/proto/schema/generated_vtproto.pb.go
… scheduler/go.mod
… scheduler/go.sum
… scheduler/plugin/plugin.go
… scheduler/test/util.go