<!--  Thanks for sending a pull request!  Here are some tips for you:

1. If this is your first time, please read our contributor guidelines: https://git.k8s.io/community/contributors/guide/first-contribution.md#your-first-contribution and developer guide https://git.k8s.io/community/contributors/devel/development.md#development-guide
2. Please label this pull request according to what type of issue you are addressing, especially if this is a release targeted pull request. For reference on required PR/issue labels, read here:
https://git.k8s.io/community/contributors/devel/sig-release/release.md#issuepr-kind-label
3. Ensure you have added or ran the appropriate tests for your PR: https://git.k8s.io/community/contributors/devel/sig-testing/testing.md
4. If you want *faster* PR reviews, read how: https://git.k8s.io/community/contributors/guide/pull-requests.md#best-practices-for-faster-reviews
5. If the PR is unfinished, see how to mark it: https://git.k8s.io/community/contributors/guide/pull-requests.md#marking-unfinished-pull-requests
-->

#### What type of PR is this?

<!--
Add one of the following kinds:
/kind bug
/kind cleanup
/kind documentation
/kind feature

Optionally add one or more of the following kinds if applicable:
/kind api-change
/kind deprecation
/kind failing-test
/kind flake
/kind regression
-->

/kind feature

#### What this PR does / why we need it:

This PR updates all k8s references from 1.27.3 to 1.30.4, which is the version targeted by the [Kubernetes Scheduler Simulator](https://github.com/kubernetes-sigs/kube-scheduler-simulator). This allows us to use the WASM plugin with the [debuggable scheduler](https://github.com/kubernetes-sigs/kube-scheduler-simulator/blob/master/simulator/docs/debuggable-scheduler.md).

#### Which issue(s) this PR fixes:
<!--
*Automatically closes linked issue when PR is merged.
Usage: `Fixes #<issue number>`, or `Fixes (paste link of issue)`.
_If PR is about `failing-tests or flakes`, please post the related issues/tests in a comment and do not use `Fixes`_*
-->

#### Special notes for your reviewer:

* Makefile, submodule, and protos updated
* Minimal version of Go updated in all go.mod files to match the kubernetes one (`1.22.0`)
* `go mod tidy` re-ran on all of them
* Code changes on the host side, but none on the guest side

#### Does this PR introduce a user-facing change?
<!--
If no, just write "NONE" in the release-note block below.
If yes, a release note is required:
Enter your extended release note in the block below. If the PR requires additional action from users switching to the new release, include the string "action required".

For more information on release notes see: https://git.k8s.io/community/contributors/guide/release-notes.md
-->
```release-note
Updated to Kubernetes 1.30.4, no guest changes needed.
```

#### What are the benchmark results of this change?
<!--
If documentation-only or trivial, just write "N/A". Otherwise, run the following:

1. `make bench-example > before.txt` on the commit prior to your changes
2. `make bench-example > after.txt` on the commit of your changes
3. Paste the output of `benchstat before.txt after.txt` as a code block.

If you haven't yet installed benchstat, it looks like this.
```
$ go install golang.org/x/perf/cmd/benchstat@latest
```

-->
```benchstat
goos: linux
goarch: arm64
pkg: sigs.k8s.io/kube-scheduler-wasm-extension/internal/e2e/scheduler
                                      │  before.txt  │             after.txt              │
                                      │    sec/op    │    sec/op     vs base              │
Example_NodeNumber/Simple/New-6         335.6m ±  4%   324.1m ±  3%  -3.42% (p=0.026 n=6)
Example_NodeNumber/Simple/Run-6         125.0µ ±  8%   118.7µ ±  7%       ~ (p=0.093 n=6)
Example_NodeNumber/Simple_Log/New-6     347.4m ±  7%   327.6m ±  4%  -5.71% (p=0.041 n=6)
Example_NodeNumber/Simple_Log/Run-6     143.3µ ± 10%   140.0µ ± 14%       ~ (p=0.485 n=6)
Example_NodeNumber/Advanced/New-6       719.2m ±  2%   721.6m ±  2%       ~ (p=0.485 n=6)
Example_NodeNumber/Advanced/Run-6       47.64µ ±  4%   48.20µ ±  9%       ~ (p=0.485 n=6)
Example_NodeNumber/Advanced_Log/New-6   716.7m ±  4%   722.0m ± 10%       ~ (p=0.485 n=6)
Example_NodeNumber/Advanced_Log/Run-6   55.63µ ±  3%   55.60µ ±  1%       ~ (p=0.699 n=6)
geomean                                 6.410m         6.294m        -1.82%
```